requirements.txt
streamlit
python-dotenv
langchain>=0.2.2
PyPDF2
chromadb
faiss-cpu
langchain_community
langchain-groq
langchain-huggingface
sentence-transformers
torch
